legacy/iot_zmq_publishers/pyproject.toml
[tool.poetry]
name = "iot-zmq-publishers"
version = "0.0.1"
description = ""
authors = ["jason a. grafft <j@grafft.co>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pyzmq = "^22.3.0"
jsonstreams = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
